requirements.txt
# Core FastAPI dependencies
fastapi==0.115.12
uvicorn[standard]==0.34.2
starlette==0.46.2

# Database
SQLAlchemy==2.0.40
aiosqlite==0.21.0

# Pydantic and validation
pydantic==2.11.4
pydantic-settings==2.3.0
pydantic_core==2.33.2

# HTTP client
requests==2.31.0

# Telegram bot
python-telegram-bot==20.8

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.9

# Development and testing
pytest==8.0.0
pytest-asyncio==0.24.0
httpx==0.27.0

# Type checking
typing-extensions==4.13.2
annotated-types==0.7.0
